ai/requirements.txt
fastapi
uvicorn
google-generativeai
matplotlib
numpy
python-dotenv
pydantic
